test(golden): 10-scenario DM-turn corpus, recorded live + replaying offline
Fold the single dm-open-door golden into a scenario-driven corpus
(corpus.golden.test.ts; dm-turn-golden.test.ts removed). Ten distinct
player intents through the real GameEngine DM loop against a seeded
"Gilded Stag" scene so the DM narrates in character rather than punting
for context:

  open-door, look-around, talk-npc, attack, skill-check, examine,
  use-item, move, social, multi-turn (2 turns).

Subagents are mocked to no-ops so each golden is a single deterministic
DM bucket. dm-skill-check captures a present_choices tool round-trip
(2 entries) and dm-multi-turn a 2-turn sequence, so the corpus exercises
tool_use/tool_result replay and cross-turn context, not just plain
narration. All 10 record live (npm run golden:record) and replay
offline with no key (npm run golden:verify / npm test).
