+++ packages/engine/src/testing/goldens/dm-attack.golden.json
@@ -1,0 +1,97 @@
+{
+  "scenario": "dm-attack",
+  "tape": {
+    "version": 1,
+    "scenario": "dm-attack",
+    "capabilities": {
+      "claude-haiku-4-5-20251001": {
+        "imageGeneration": false
+      }
+    },
+    "entries": [
+      {
+        "kind": "chat",
+        "bucket": "dm",
+        "ordinal": 0,
+        "request": {
+          "model": "claude-haiku-4-5-20251001",
+          "messageCount": 1,
+          "systemHash": "9d270259c954",
+          "tools": [
+            "advance_turn",
+            "alarm",
+            "deck",
+            "describe_entity_type",
+            "detect_orphans",
+            "dm_notes",
+            "end_combat",
+            "enter_ooc",
+            "entity",
+            "find_schema_drift",
+            "list_entity_types",
+            "manage_objectives",
+            "map",
+            "map_entity",
+            "map_query",
+            "modify_initiative",
+            "present_choices",
+            "promote_character",
+            "resolve_turn",
+            "roll_dice",
+            "scene_transition",
+            "scribe",
+            "search_campaign",
+            "search_content",
+            "session_end",
+            "set_display_resources",
+            "set_resource_values",
+            "start_combat",
+            "style_scene",
+            "switch_player",
+            "time",
+            "update_modeline",
+            "validate_entity"
+          ],
+          "lastMessagePreview": "<context> ## Current State PCs: Aldric [length] Terminal: 80 cols × 16 visible rows. [stats] Turn: Aldric </context> [Al"
+        },
+        "result": {
+          "text": "The <color=#20b2aa>oak door</color> splinters at the hinges with a <b>crack</b> that echoes through the tavern. Rain and wind gust through the common room as everyone's head snaps toward you.\n\nMella's face goes pale. \"What in the—\"\n\nThe hooded figure <i>moves</i>—swift, purposeful—shoving back from the hearth and bolting for the tavern's front exit without a word. In that split second you catch a glimpse of a scarred jawline before the hood swallows their face again.\n\nBelow the cellar door: <b>darkness</b>. A stone stairwell descends into black. The air that rises from it is cold and carries the metallic reek of old blood.\n\nMella grabs the bar with both hands, knuckles white. \"You fool,\" she hisses. \"You don't know what you've done.\"\n\nThe front door slams. The hooded figure is gone into the rain.\n\nWhat do you do?",
+          "toolCalls": [],
+          "usage": {
+            "inputTokens": 393,
+            "outputTokens": 240,
+            "cacheReadTokens": 9506,
+            "cacheCreationTokens": 0,
+            "reasoningTokens": 0
+          },
+          "stopReason": "end",
+          "assistantContent": [
+            {
+              "type": "text",
+              "text": "The <color=#20b2aa>oak door</color> splinters at the hinges with a <b>crack</b> that echoes through the tavern. Rain and wind gust through the common room as everyone's head snaps toward you.\n\nMella's face goes pale. \"What in the—\"\n\nThe hooded figure <i>moves</i>—swift, purposeful—shoving back from the hearth and bolting for the tavern's front exit without a word. In that split second you catch a glimpse of a scarred jawline before the hood swallows their face again.\n\nBelow the cellar door: <b>darkness</b>. A stone stairwell descends into black. The air that rises from it is cold and carries the metallic reek of old blood.\n\nMella grabs the bar with both hands, knuckles white. \"You fool,\" she hisses. \"You don't know what you've done.\"\n\nThe front door slams. The hooded figure is gone into the rain.\n\nWhat do you do?"
+            }
+          ]
+        },
+        "streamDeltas": [
+          "The",
+          " <color=#20b2aa>oak door</color> splinters at the hinges with a <b>crack",
+          "</b> that echoes through the tavern. Rain and wind gust through the common room as everyone's",
+          " head snaps toward you.\n\nMella's face goes pale. \"",
+          "What in the—\"\n\nThe hooded figure <i>moves</i>",
+          "—swift, purposeful—shoving back from the hear",
+          "th and bolting for the tavern's front exit without a word. In",
+          " that split second you catch a glimpse of a scarred jawline before the hood",
+          " swallows their face again.\n\nBelow the cellar door: <b>darkness</b>. A",
+          " stone stairwell descends into black. The air that rises",
+          " from it is cold and carries the metallic reek of old blood.\n\nMella grabs the",
+          " bar with both hands, knuckles white. \"You fool,\" she hisses. \"You don't know what you've",
+          " done.\"\n\nThe front door slams. The hooded figure is gone into the rain.",
+          "\n\nWhat do you do?"
+        ]
+      }
+    ]
+  },
+  "expectedNarrative": [
+    "The <color=#20b2aa>oak door</color> splinters at the hinges with a <b>crack</b> that echoes through the tavern. Rain and wind gust through the common room as everyone's head snaps toward you.\n\nMella's face goes pale. \"What in the—\"\n\nThe hooded figure <i>moves</i>—swift, purposeful—shoving back from the hearth and bolting for the tavern's front exit without a word. In that split second you catch a glimpse of a scarred jawline before the hood swallows their face again.\n\nBelow the cellar door: <b>darkness</b>. A stone stairwell descends into black. The air that rises from it is cold and carries the metallic reek of old blood.\n\nMella grabs the bar with both hands, knuckles white. \"You fool,\" she hisses. \"You don't know what you've done.\"\n\nThe front door slams. The hooded figure is gone into the rain.\n\nWhat do you do?"
+  ]
+}
