+++ packages/engine/src/testing/goldens/dm-social.golden.json
@@ -1,0 +1,99 @@
+{
+  "scenario": "dm-social",
+  "tape": {
+    "version": 1,
+    "scenario": "dm-social",
+    "capabilities": {
+      "claude-haiku-4-5-20251001": {
+        "imageGeneration": false
+      }
+    },
+    "entries": [
+      {
+        "kind": "chat",
+        "bucket": "dm",
+        "ordinal": 0,
+        "request": {
+          "model": "claude-haiku-4-5-20251001",
+          "messageCount": 1,
+          "systemHash": "9d270259c954",
+          "tools": [
+            "advance_turn",
+            "alarm",
+            "deck",
+            "describe_entity_type",
+            "detect_orphans",
+            "dm_notes",
+            "end_combat",
+            "enter_ooc",
+            "entity",
+            "find_schema_drift",
+            "list_entity_types",
+            "manage_objectives",
+            "map",
+            "map_entity",
+            "map_query",
+            "modify_initiative",
+            "present_choices",
+            "promote_character",
+            "resolve_turn",
+            "roll_dice",
+            "scene_transition",
+            "scribe",
+            "search_campaign",
+            "search_content",
+            "session_end",
+            "set_display_resources",
+            "set_resource_values",
+            "start_combat",
+            "style_scene",
+            "switch_player",
+            "time",
+            "update_modeline",
+            "validate_entity"
+          ],
+          "lastMessagePreview": "<context> ## Current State PCs: Aldric [length] Terminal: 80 cols × 16 visible rows. [stats] Turn: Aldric </context> [Al"
+        },
+        "result": {
+          "text": "You weave between the crowded tables—locals hunched over their ales, avoiding eye contact—and plant yourself across from the hooded figure with an easy smile.\n\n\"Rough night out there,\" you say, catching the barkeep's eye and raising two fingers. \"Let me buy you another. What's your poison?\"\n\nThe figure doesn't look up immediately. When they do, you glimpse a weathered face, jaw tight, eyes the color of storm clouds. <i>\"I'm fine,\"</i> comes the response—flat, unwelcoming.\n\nMella arrives with two fresh tankards. She sets them down with more force than necessary, her gaze flicking to the hooded one before settling back on you. \"That'll be four coppers,\" she says pointedly.\n\nThe figure shifts. They seem to be weighing something—whether to ignore you entirely or risk a conversation. After a long moment: <i>\"What do you want to know?\"</i> Not friendly. Not hostile either. Cautious.\n\n<i>The rain drums harder against the shutters. The tavern smells of woodsmoke and damp wool.</i>\n\nWhat do you ask them?",
+          "toolCalls": [],
+          "usage": {
+            "inputTokens": 394,
+            "outputTokens": 261,
+            "cacheReadTokens": 9506,
+            "cacheCreationTokens": 0,
+            "reasoningTokens": 0
+          },
+          "stopReason": "end",
+          "assistantContent": [
+            {
+              "type": "text",
+              "text": "You weave between the crowded tables—locals hunched over their ales, avoiding eye contact—and plant yourself across from the hooded figure with an easy smile.\n\n\"Rough night out there,\" you say, catching the barkeep's eye and raising two fingers. \"Let me buy you another. What's your poison?\"\n\nThe figure doesn't look up immediately. When they do, you glimpse a weathered face, jaw tight, eyes the color of storm clouds. <i>\"I'm fine,\"</i> comes the response—flat, unwelcoming.\n\nMella arrives with two fresh tankards. She sets them down with more force than necessary, her gaze flicking to the hooded one before settling back on you. \"That'll be four coppers,\" she says pointedly.\n\nThe figure shifts. They seem to be weighing something—whether to ignore you entirely or risk a conversation. After a long moment: <i>\"What do you want to know?\"</i> Not friendly. Not hostile either. Cautious.\n\n<i>The rain drums harder against the shutters. The tavern smells of woodsmoke and damp wool.</i>\n\nWhat do you ask them?"
+            }
+          ]
+        },
+        "streamDeltas": [
+          "You",
+          " weave between the crowded tables—locals hunched over their ales, avoiding",
+          " eye contact—and plant yourself across from the hooded figure with an easy smile.",
+          "\n\n\"Rough night out there,\" you say, catching the barkeep's eye and",
+          " raising two fingers. \"Let me buy you another. What's your poison",
+          "?\"\n\nThe figure doesn't look up immediately. When they do, you glimpse a weather",
+          "ed face, jaw tight, eyes the color of storm clouds. <i",
+          ">\"I'm fine,\"</i> comes the response—flat, unw",
+          "elcoming.\n\nMella arrives with two fresh tankards. She",
+          " sets them down with more force than necessary, her gaze flicking to the hooded one before",
+          " settling back on you. \"That'll be four coppers,\" she says pointedly.\n\nThe figure shifts",
+          ". They seem to be weighing something—whether to ignore you",
+          " entirely or risk a conversation. After a long moment: <i>\"What do you want to",
+          " know?\"</i> Not friendly. Not hostile either. Caut",
+          "ious.\n\n<i>The rain drums harder against the shutters. The tavern sm",
+          "ells of woodsmoke and damp wool.</i>\n\nWhat do you ask them?"
+        ]
+      }
+    ]
+  },
+  "expectedNarrative": [
+    "You weave between the crowded tables—locals hunched over their ales, avoiding eye contact—and plant yourself across from the hooded figure with an easy smile.\n\n\"Rough night out there,\" you say, catching the barkeep's eye and raising two fingers. \"Let me buy you another. What's your poison?\"\n\nThe figure doesn't look up immediately. When they do, you glimpse a weathered face, jaw tight, eyes the color of storm clouds. <i>\"I'm fine,\"</i> comes the response—flat, unwelcoming.\n\nMella arrives with two fresh tankards. She sets them down with more force than necessary, her gaze flicking to the hooded one before settling back on you. \"That'll be four coppers,\" she says pointedly.\n\nThe figure shifts. They seem to be weighing something—whether to ignore you entirely or risk a conversation. After a long moment: <i>\"What do you want to know?\"</i> Not friendly. Not hostile either. Cautious.\n\n<i>The rain drums harder against the shutters. The tavern smells of woodsmoke and damp wool.</i>\n\nWhat do you ask them?"
+  ]
+}
